Extract Scroll class from queryService.js
